backend/auth_service/pyproject.toml
[project]
name = "auth-service"
version = "0.1.0"
description = ""
authors = [
    {name = "IvanZaycev0717",email = "ivzaycev0717@yandex.ru"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "colorlog (>=6.9.0,<7.0.0)",
    "pydantic-settings (>=2.11.0,<3.0.0)",
    "grpcio (>=1.75.1,<2.0.0)",
    "grpcio-tools (>=1.75.1,<2.0.0)",
    "protobuf (>=6.32.1,<7.0.0)",
    "pymongo (>=4.15.2,<5.0.0)",
    "bcrypt (>=5.0.0,<6.0.0)",
    "pydantic[email] (>=2.11.10,<3.0.0)",
    "grpcio-health-checking (>=1.75.1,<2.0.0)",
    "grpcio-reflection (>=1.75.1,<2.0.0)",
    "jwt (==1.4.0)",
    "mongomock (>=4.3.0,<5.0.0)",
    "aiokafka (>=0.12.0,<0.13.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.2"
pytest-asyncio = "^1.2.0"
isort = "^6.1.0"
ruff = "^0.14.0"

[tool.pytest.ini_options]
asyncio_mode = "auto"
filterwarnings = [
    "ignore::DeprecationWarning",
]
addopts = [
    "--import-mode=importlib",
]
pythonpath = ["src"]

